improve the code
